requirements.txt
flask
flask-login
flask-script
flask-sqlalchemy
WTForms